+++ qml/CibaLogin.qml
@@ -1,13 +1,18 @@
 import QtQuick 2.15
 import QtQuick.Controls 2.15
 import QtQuick.Layouts 1.3
-
 import MatrixClient 1.0
+import QmlInterface 1.0
 
 CustomPage {
     id: loginPage
     width: parent.width
     spacing: 10
+
+    Validator{
+        id: validator
+    }
+
     ColumnLayout{
         id: inputLayout
         anchors.centerIn: parent
@@ -17,7 +22,20 @@ CustomPage {
             Layout.leftMargin: 50
             Layout.rightMargin: 50
             Layout.fillWidth: true
-            placeholderText: qsTr("Email")
+            placeholderText: qsTr("ID")
+            Keys.onReturnPressed: loginButton.gotoLogin()
+            Keys.onEnterPressed: loginButton.gotoLogin()
+        }
+
+        TextField {
+            id: matrixServerText
+            // text: "https://matrix.pantherx.org"
+            Layout.leftMargin: 50
+            Layout.rightMargin: 50
+            Layout.fillWidth: true
+            placeholderText: "Matrix Server (e.g.: " + QmlInterface.defaultMatrixServer() + ")"
+            text: QmlInterface.isSetServerAsDefault()?QmlInterface.defaultMatrixServer():""
+            validator: validator.matrixServerRegex()
             Keys.onReturnPressed: loginButton.gotoLogin()
             Keys.onEnterPressed: loginButton.gotoLogin()
         }
@@ -36,8 +54,9 @@ CustomPage {
                 anchors.verticalCenter: parent.verticalCenter
                 function gotoLogin(){
                     loginButton.enabled= false;
+                    matrixServerText.text = validator.checkMatrixServerUrl(matrixServerText.text)
                     MatrixClient.loginWithCiba(String(userIdText.text),
-                                            String("https://matrix.pantherx.dev"))
+                                               String(matrixServerText.text))
                 }
 
                 onClicked: gotoLogin()