+++ qml/Login.qml
@@ -1,11 +1,16 @@
 import QtQuick 2.15
 import QtQuick.Controls 2.15
 import QtQuick.Layouts 1.3
-
 import MatrixClient 1.0
+import QmlInterface 1.0
 
 CustomPage {
     id: loginPage
+
+    Validator{
+        id: validator
+    }
+
     ColumnLayout{
         id: inputLayout
         anchors.centerIn: parent
@@ -15,8 +20,9 @@ CustomPage {
             Layout.leftMargin: 50
             Layout.rightMargin: 50
             Layout.fillWidth: true
-            // text: "hamzeh_test04"
-            placeholderText: qsTr("User ID")
+            validator: validator.userIdRegex()
+            // text: "@hamzeh_test05:pantherx.org"
+            placeholderText: "User ID" + (QmlInterface.defaultUserIdFormat()?" (e.g.: " + QmlInterface.defaultUserIdFormat() + ")" : "")
             Keys.onReturnPressed: loginButton.gotoLogin()
             Keys.onEnterPressed: loginButton.gotoLogin()
         }
@@ -24,11 +30,24 @@ CustomPage {
         TextField {
             id: passwordText
             echoMode: TextInput.Password
-            // text: "tingly-headdress-earthlike"
+            // text: "riverbed-judiciary-sworn"
             Layout.leftMargin: 50
             Layout.rightMargin: 50
             Layout.fillWidth: true
             placeholderText: qsTr("Password")
+            Keys.onReturnPressed: loginButton.gotoLogin()
+            Keys.onEnterPressed: loginButton.gotoLogin()
+        }
+
+        TextField {
+            id: matrixServerText
+            // text: "https://matrix.pantherx.org"
+            Layout.leftMargin: 50
+            Layout.rightMargin: 50
+            Layout.fillWidth: true
+            validator: validator.matrixServerRegex()
+            placeholderText: "Matrix Server (e.g.: " + QmlInterface.defaultMatrixServer() + ")"
+            text: QmlInterface.isSetServerAsDefault()?QmlInterface.defaultMatrixServer():""
             Keys.onReturnPressed: loginButton.gotoLogin()
             Keys.onEnterPressed: loginButton.gotoLogin()
         }
@@ -39,10 +58,11 @@ CustomPage {
             Layout.alignment: Qt.AlignHCenter
             function gotoLogin(){
                 loginButton.enabled= false;
+                matrixServerText.text = validator.checkMatrixServerUrl(matrixServerText.text)
                 MatrixClient.loginWithPassword(String("matrix_client_application"),
-                                         String("@" + userIdText.text + ":pantherx.org"),
+                                         String(userIdText.text),
                                          String(passwordText.text),
-                                         String("https://matrix.pantherx.org"))
+                                         String(matrixServerText.text))
             }
 
             onClicked: gotoLogin()