+++ qml/RoomList.qml
@@ -4,6 +4,7 @@ import QtQuick.Controls 2.5
 
 import MatrixClient 1.0
 import Rooms 1.0
+import GlobalObject 1.0
 
 import "device-verification"
 
@@ -15,7 +16,7 @@ CustomPage {
     ListView {
         id: roomListView
         anchors.fill: parent
-        spacing: 10
+        spacing: 0
         anchors.margins: 10
         flickableDirection: Flickable.VerticalFlick
         boundsBehavior: Flickable.StopAtBounds
@@ -23,6 +24,7 @@ CustomPage {
         model: Rooms
         delegate:RoomDelegate{}
     }
+
     DirectChatDialog{
         id:directChat
         x: (qmlLibRoot.width - width) / 2
@@ -35,6 +37,7 @@ CustomPage {
         width: height 
         x: parent.width - width -10
         y: parent.height - height -10
+        palette.button: GlobalObject.colors.alternateBase
         font.pointSize: 15            
         text: "+"
         onClicked: directChat.open()