+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 
+## [0.0.8]
+### Changed
+
+* Update icons.
+* Update QML style to match with OS Theme.
+* Get Matrix Server address from user input in Login and CIBA Login pages.
+* Set validator on Matrix Server and USER ID in the Forms.
+
+
 ## [0.0.7]
 ### Changed